add reindexing of failed claim-sets
claim-sets which could not be indexed because the attr-factory script
failed will be reindexd later by evr-attr-index.

this will be essential when later the sync between two
evr-glalier-storage instances is implemented. because it might happend
that a claim-set which contains a file is synced before the actual
file contents is synced.