fix flaky fs integration test failures when waiting for new-category
prior to this patch the seconds fraction part of the claim-set
creation dates was always set to 999999. that led to the situation
that sometimes the creation date was in the future. so the watch
trigger towards evr-fs updated the files based on the seed but got the
seeds old data.