use flock in evr-glacier-storage for db locking
evr-glacier-storage used the existence of a lock file to detect if
another process is already using the glacier buckets.

this approach did unexpected power loss not handle very well. the lock
file was not deleted and evr-glacier-storage refuses to start on
reboot.

the lock file is now locked using flock after creation to put an
exclusive lock onto the glacier's lock file.