fix bucket dir walking in musl
the bucket dir walking depends on the readdir(…) function call and
the value it's setting to errno. the code within the bucket walking
set errno to 0 initially and did then not touch errno anymore while
walking over the files within the buckets directory. on a successful
run errno stood 0 until the end of the readdir stream was reached.

in musl errno is set to 22 during a log_debug(…) call. so resetting
errno once to 0 before the walk was not enough.