move integration test execution to Makefiles
from test.sh. that should execute the integration tests before the
podman images are built. the integration tests should be executed
before podman because the podman builds also execute the integration
test. the feedback loop is much smaller if the integration tests fail
on the host system than when after building an podman image and
running the container.