force index.db path within evr-backup
evr-backup starts an evr-glacier-storage server within the backup
target directory. the path of the target index.db was not explicitly
specified. that's why a custom evr-glacier-storage.conf file could
alter the location. that made the evr-backup script overwrite that
location.

with this fix the index.db of the backup target lies always within the
backup target directory.